+++ components/Button/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shallot-ui/component-button
 
+## 3.0.0-beta.11
+
+### Patch Changes
+
+- Design review on components
+
 ## 3.0.0-beta.10
 
 ### Minor Changes